+++ test/dom/EventTrackingServiceDOM.spec.ts
@@ -23,6 +23,7 @@ const fetchMocker = createFetchMock(vi);
 
 // Use console logs while testing. Could be an env variable in the future.
 const logOutcomes: boolean = false;
+const raiseExceptions: boolean = true;
 
 describe(`Testing the EventTrackingService using JSDOM`, () => {
   beforeAll(() => {
@@ -90,7 +91,7 @@ describe(`Testing the EventTrackingService using JSDOM`, () => {
 
     // Runs in a browser-like environment
     const error = await getError<Error>(async () => {
-      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes);
+      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes, raiseExceptions);
     });
 
     expect(error).not.toBeInstanceOf(NoErrorThrownError);
@@ -112,7 +113,7 @@ describe(`Testing the EventTrackingService using JSDOM`, () => {
 
     // Runs in a browser-like environment
     const error = await getError<Error>(async () => {
-      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes);
+      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes, raiseExceptions);
     });
 
     expect(error).not.toBeInstanceOf(NoErrorThrownError);
@@ -136,7 +137,7 @@ describe(`Testing the EventTrackingService using JSDOM`, () => {
 
     // Runs in a browser-like environment
     const error = await getError<Error>(async () => {
-      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes);
+      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes, raiseExceptions);
     });
 
     expect(error).not.toBeInstanceOf(NoErrorThrownError);
@@ -171,7 +172,7 @@ describe(`Testing the EventTrackingService using JSDOM`, () => {
 
     // Runs in a browser-like environment
     const error = await getError<Error>(async () => {
-      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes);
+      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes, raiseExceptions);
     });
 
     expect(error).toBeInstanceOf(NoErrorThrownError);
@@ -190,7 +191,7 @@ describe(`Testing the EventTrackingService using JSDOM`, () => {
 
     // Runs in a browser-like environment
     const error = await getError<Error>(async () => {
-      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes);
+      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes, raiseExceptions);
     });
 
     expect(error).not.toBeInstanceOf(NoErrorThrownError);
@@ -209,7 +210,7 @@ describe(`Testing the EventTrackingService using JSDOM`, () => {
 
     // Runs in a browser-like environment
     const error = await getError<Error>(async () => {
-      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes);
+      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes, raiseExceptions);
     });
 
     expect(error).not.toBeInstanceOf(NoErrorThrownError);
@@ -226,12 +227,26 @@ describe(`Testing the EventTrackingService using JSDOM`, () => {
 
     // Runs in a browser-like environment
     const error = await getError<Error>(async () => {
-      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes);
+      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes, raiseExceptions);
     });
 
     expect(error).not.toBeInstanceOf(NoErrorThrownError);
     expect(fetchMocker.requests().length).toBe(1); // Does not retry on failure
     expect(error.message).toBe("Error interfacing with Mini Digital.");
+  });
+
+  it("Does not throw exceptions when raising exceptions is not true", async () => {
+    fetchMocker.once((request) => {
+      describeRequest(request, logOutcomes);
+      throw new Error("Simulating throwing an error");
+    });
+
+    // Runs in a browser-like environment
+    const error = await getError<Error>(async () => {
+      await EventTrackingService.postEvent(VALID_ANONYMOUS_EVENT, logOutcomes);
+    });
+
+    expect(error).toBeInstanceOf(NoErrorThrownError);
   });
 
   it("Doesn't send a request when config.pauseTracking is true", async () => {